refactor: hoist pageText + AI block-range helpers into core
getPageTextFromLayout and the AI-edit block-range helpers (resolveFolioAIBlockRange, clampRangeToDocSize) were duplicated verbatim in both adapters (each with its own copy + tests, and comments documenting the forced duplication). Move them to framework-neutral core modules (@stll/folio-core/paged-layout/pageText, @stll/folio-core/ai-edits/blockRange) with a single shared test suite; React and Vue now import from core. Arch tests (react-free-core, layer-boundaries, model-purity) stay green.
